+++ Student-Management-System/src/main/java/com/easyWay/Student_Management_System/ServiceImpl/StudentFeesServiceImpl.java
@@ -45,9 +45,9 @@ public class StudentFeesServiceImpl implements StudentFeesService {
             studentFeeInfo.setFee(studentFees.getFee());
             studentFeeInfo.setStudentInfo(studentInfo.get());
             studentFeesInfoRepo.save(studentFeeInfo);
-            int remaining = studentInfo.get().getRemainingFees();
+            float remaining = studentInfo.get().getRemainingFees();
             remaining = remaining- studentFees.getFee();
-            studentInfo.get().setRemainingFees(remaining);
+            studentInfo.get().setRemainingFees((int)remaining);
             studentFeeInfo.setSchoolCode(claimService.getLoggedInUserSchoolCode());
             studentInfoRepo.save(studentInfo.get());
         }
@@ -68,9 +68,9 @@ public class StudentFeesServiceImpl implements StudentFeesService {
         if (ObjectUtils.isEmpty(entity)){
             throw new BadRequestException("Not data found");
         }
-        int rem = dto.getFee()-entity.getFee();// 900 -899 = 1
+        float rem = dto.getFee()-entity.getFee();// 900 -899 = 1
         int oldRem = entity.getStudentInfo().getRemainingFees();//280
-        entity.getStudentInfo().setRemainingFees(oldRem-rem);
+        entity.getStudentInfo().setRemainingFees((int) (oldRem-rem));
         entity.setFee(dto.getFee());
         studentFeesInfoRepo.save(entity);
         return "data edit successfully";
